Refactor account operations into account go file
